virt-who config cases add create & delete to virtwho_config fixture for CLI/API/UI (#11245)
* virt-who config cases add create & delete to virtwho_config fixture

* Update tests/foreman/virtwho/api/test_esx.py


* Update tests/foreman/virtwho/api/test_libvirt_sca.py


* Update test_esx.py

change the fixture name to "delete_host"

---------

(cherry picked from commit be592c3e59927a05cb0761e76a141c589926e9d4)
